Don't memoize document changes
In certain save situations this will memoize to false when you really
have changes. For instance, an embedded document is set and calls save
on it's parent (our document in question) then the document proceeds to
be updated. Changes are memoized as being nonexistent and history record
is not created.